쇼핑몰 패키지 이름 수정
