--- shoppingmall/src/main/java/ShoppingMall/controller/controller.java
@@ -1,5 +1,0 @@
-package ShoppingMall.controller;
-
-public class controller {
-	
-}
