--- shoppingmall/src/main/java/ShoppingMall/mybatis/mybatis-config.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE configuration 
-PUBLIC "-//mybatis.org//DTD Config 3.0//EN" "https://mybatis.org/dtd/mybatis-3-config.dtd">
-
-<configuration>
-	<typeAliases>
-		<typeAlias type="ShoppingMall.domain.Status"		alias="Status"/>
-		<typeAlias type="ShoppingMall.domain.Qna"			alias="Qna"/>
-		<typeAlias type="ShoppingMall.domain.Returns"		alias="Returns"/>
-		<typeAlias type="ShoppingMall.domain.Delivery"		alias="Delivery"/>
-		<typeAlias type="ShoppingMall.domain.Member"		alias="Member"/>
-		<typeAlias type="ShoppingMall.domain.OrderSummary"	alias="OrderSummary"/>
-
-	</typeAliases>
-	
-	<environments default="development">
-		<environment id="development">
-			<transactionManager type="JDBC" />
-			<dataSource type="JNDI">
-				<property name="data_source" value="java:comp/env/jndi/mysql" />
-			</dataSource>
-		</environment>
-	</environments>
-	<mappers>
-		<mapper resource="ShoppingMall/mybatis/StatusMapper.xml" />
-		<mapper resource="ShoppingMall/mybatis/QnaMapper.xml" />
-		<mapper resource="ShoppingMall/mybatis/ReturnsMapper.xml" />
-		<mapper resource="ShoppingMall/mybatis/DeliveryMapper.xml" />
-		<mapper resource="ShoppingMall/mybatis/MemberMapper.xml" />
-		<mapper resource="ShoppingMall/mybatis/OrderSummaryMapper.xml" />
-	</mappers>
-</configuration>
