+++ shoppingmall/src/main/java/shoppingmall/controller/controller.java
@@ -1,0 +1,5 @@
+package shoppingmall.controller;
+
+public class controller {
+	
+}
