+++ shoppingmall/src/main/java/shoppingmall/domain/Delivery.java
@@ -1,4 +1,4 @@
-package ShoppingMall.domain;
+package shoppingmall.domain;
 
 import javax.persistence.OneToOne;
