+++ shoppingmall/src/main/java/shoppingmall/domain/Qna.java
@@ -1,4 +1,4 @@
-package ShoppingMall.domain;
+package shoppingmall.domain;
 
 import lombok.Data;
