+++ shoppingmall/src/main/java/shoppingmall/domain/Status.java
@@ -1,4 +1,4 @@
-package ShoppingMall.domain;
+package shoppingmall.domain;
 
 import lombok.Data;
