+++ shoppingmall/src/main/java/shoppingmall/mybatis/QnaMapper.xml
