+++ shoppingmall/src/main/java/shoppingmall/mybatis/StatusMapper.xml
